+++ CHANGELOG.md
@@ -6,6 +6,7 @@
     - https://speakerdeck.com/alli83/symfony-messenger-et-ses-messages-a-la-queleuleu-dot-dot-dot-et-sil-etait-temps-de-grouper
     - https://wolfgang-klinger.medium.com/how-to-handle-messages-in-batches-with-symfony-messenger-c91b5aa1c8b1
     - https://github.com/wazum/symfony-messenger-batch
+- YFlow rework
 
 ## v1.2.0